Zadanie 6.4 v2

+++ kodilla-testing/src/main/java/com/kodilla/testing/shape/Shape.java
@@ -3,5 +3,5 @@ package com.kodilla.testing.shape;
 interface Shape {
 
     public String getShapeName();
-    public int getField();
+    public double getField();
 }
